requirements.txt
Flask
Flask-CORS
gunicorn
requests
openpyxl
psycopg2-binary